Wishlist functionality

+++ MCIA/MCIA/src/WishList.cpp
@@ -1,6 +1,6 @@
 #include "../include/WishList.h"
 
-WishList::WishList(const uint16_t user_id, const uint16_t movie_id)
+WishList::WishList(const uint16_t user_id, const uint32_t movie_id)
 	:m_user_id(user_id)
 	,m_movie_id(movie_id)
 {
@@ -11,7 +11,7 @@ uint16_t WishList::GetUserId() const
 	return m_user_id;
 }
 
-uint16_t WishList::GetMovieId() const
+uint32_t WishList::GetMovieId() const
 {
 	return m_movie_id;
 }
@@ -21,7 +21,7 @@ void WishList::SetUserId(const uint16_t user_id)
 	m_user_id = user_id;
 }
 
-void WishList::SetMovieId(const uint16_t movie_id)
+void WishList::SetMovieId(const uint32_t movie_id)
 {
 	m_movie_id = movie_id;
 }
